requirements.txt
aiogram
python-dotenv
pytest